requirements.txt
numpy
pandas
loguru
python-binance